requirements.txt
Django==1.10.6
Pygments==2.2.0
decorator==4.0.11
djangorestframework==3.6.2
ipython==5.3.0
mock==2.0.0
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
six==1.10.0
